embedding/requirements.txt
openai==1.12.0
python-dotenv==1.0.0
pymysql==1.1.0
kiwipiepy>=0.17.0
keybert>=0.7.0
sentence-transformers>=2.2.2
pymilvus==2.3.4
kss>=2.5.2
tqdm>4
numpy>=1.24.0
pandas>=1.2.4
scikit-learn>=0.22.2
torch>=1.11.0
transformers>=4.41.0 